# Improve efficiency of the scan for dirty nodes