--- a/go/database/mpt/forest.go
+++ b/go/database/mpt/forest.go
@@ -523,10 +523,7 @@ func (s *Forest) Flush() error {
 	// Get snapshot of set of dirty Node IDs.
 	ids := make([]NodeId, 0, 1<<16)
 	s.nodeCache.ForEach(func(id NodeId, node *shared.Shared[Node]) {
-		handle := node.GetViewHandle()
-		dirty := handle.Get().IsDirty()
-		handle.Release()
-		if dirty {
+		if node.GetUnprotected().IsDirty() {
 			ids = append(ids, id)
 		}
 	})
@@ -553,6 +550,10 @@ func (s *Forest) flushDirtyIds(ids []NodeId) error {
 		if present {
 			handle := node.GetWriteHandle()
 			node := handle.Get()
+			if !node.IsDirty() {
+				handle.Release()
+				continue
+			}
 			err := s.flushNode(id, node)
 			if err == nil {
 				node.MarkClean()