--- a/go/database/mpt/forest.go
+++ b/go/database/mpt/forest.go
@@ -911,7 +911,7 @@ func (s *Forest) createAccount() (NodeReference, shared.WriteHandle[Node], error
 	instance, present := s.addToCache(&ref, shared.MakeShared[Node](node))
 	if present {
 		write := instance.GetWriteHandle()
-		*write.Get().(*AccountNode) = *node
+		write.Get().Reset()
 		write.Release()
 	}
 	return ref, instance.GetWriteHandle(), err
@@ -927,7 +927,7 @@ func (s *Forest) createBranch() (NodeReference, shared.WriteHandle[Node], error)
 	instance, present := s.addToCache(&ref, shared.MakeShared[Node](node))
 	if present {
 		write := instance.GetWriteHandle()
-		*write.Get().(*BranchNode) = *node
+		write.Get().Reset()
 		write.Release()
 	}
 	return ref, instance.GetWriteHandle(), err
@@ -943,7 +943,7 @@ func (s *Forest) createExtension() (NodeReference, shared.WriteHandle[Node], err
 	instance, present := s.addToCache(&ref, shared.MakeShared[Node](node))
 	if present {
 		write := instance.GetWriteHandle()
-		*write.Get().(*ExtensionNode) = *node
+		write.Get().Reset()
 		write.Release()
 	}
 	return ref, instance.GetWriteHandle(), err
@@ -959,7 +959,7 @@ func (s *Forest) createValue() (NodeReference, shared.WriteHandle[Node], error) 
 	instance, present := s.addToCache(&ref, shared.MakeShared[Node](node))
 	if present {
 		write := instance.GetWriteHandle()
-		*write.Get().(*ValueNode) = *node
+		write.Get().Reset()
 		write.Release()
 	}
 	return ref, instance.GetWriteHandle(), err