This PR applies some performance improvement on the scan for dirty nodes during flush operations. The main change is to track the dirty-state of nodes using atomic operations to facilitate checking for dirty nodes without the need of acquiring node-access permissions.

This reduces the time it takes to search for dirty nodes by more than 60%:
```
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Carmen/go/database/mpt
cpu: Intel(R) Core(TM) i7-6600U CPU @ 2.60GHz
                                              │  before.log  │              after.log               │
                                              │    sec/op    │    sec/op     vs base                │
NodeFlusher_CollectDirtyNodes/size=1000-4       33.796µ ± 4%   5.372µ ±  3%  -84.10% (p=0.000 n=10)
NodeFlusher_CollectDirtyNodes/size=10000-4      335.45µ ± 3%   54.32µ ±  7%  -83.81% (p=0.000 n=10)
NodeFlusher_CollectDirtyNodes/size=100000-4      3.891m ± 1%   1.462m ±  3%  -62.41% (p=0.000 n=10)
NodeFlusher_CollectDirtyNodes/size=1000000-4     39.50m ± 3%   14.70m ±  2%  -62.78% (p=0.000 n=10)
NodeFlusher_CollectDirtyNodes/size=10000000-4    399.0m ± 4%   142.6m ± 30%  -64.25% (p=0.000 n=10)
geomean                                          3.702m        978.0µ        -73.58%
```

### Notes for Reviewers
By introducing atomic operations on the `clean` flag in the `nodeBase` assignment operations on the node level like the one done in the `Forest` implementation (e.g. `*write.Get().(*AccountNode) = *node`) introduce a race condition since the reset of the `clean` flag is not atomic in this case. Thus, a new `Reset` function needed to be introduced, performing the `clean` flag reset using an atomic operation.